+++ brewtils/rest/easy_client.py
@@ -1159,17 +1159,18 @@ class EasyClient(object):
             return False, metadata_json
 
     @wrap_response(parse_method="parse_topic", parse_many=False, default_exc=FetchError)
-    def get_topic(self, topic_id):
+    def get_topic(self, topic_id=None, topic_name=None):
         """Get a topic
 
         Args:
-            topic_id: Topic id
+            topic_id (Optional[str]): Topic id
+            topic_name (Optional[str]): Topic name
 
         Returns:
             The Topic
 
         """
-        return self.client.get_topic(topic_id)
+        return self.client.get_topic(topic_id=topic_id, topic_name=topic_name)
 
     @wrap_response(parse_method="parse_topic", parse_many=True, default_exc=FetchError)
     def get_topics(self):
@@ -1195,11 +1196,12 @@ class EasyClient(object):
         return self.client.post_topics(SchemaParser.serialize_topic(topic))
 
     @wrap_response(return_boolean=True, raise_404=True)
-    def remove_topic(self, topic_id):
+    def remove_topic(self, topic_id=None, topic_name=None):
         """Remove a unique Topic
 
         Args:
-            topic_id: Topic id
+            topic_id (Optional[str]): Topic id
+            topic_name (Optional[str]): Topic name
 
         Returns:
             bool: True if removal was successful
@@ -1208,14 +1210,15 @@ class EasyClient(object):
             NotFoundError: Couldn't find a Topic matching given parameters
 
         """
-        return self.client.delete_topic(topic_id)
+        return self.client.delete_topic(topic_id=topic_id, topic_name=topic_name)
 
     @wrap_response(parse_method="parse_topic", parse_many=False, default_exc=SaveError)
-    def update_topic(self, topic_id, add=None, remove=None):
+    def update_topic(self, topic_id=None, topic_name=None, add=None, remove=None):
         """Update a Topic
 
         Args:
-            topic_id (str): The Topic ID
+            topic_id (Optional[str]): Topic id
+            topic_name (Optional[str]): Topic name
             add (Optional[str]): Add subscriber
             remove (Optional[str]): Remove subscriber
 
@@ -1231,5 +1234,5 @@ class EasyClient(object):
             operations.append(PatchOperation("remove", value=remove))
 
         return self.client.patch_topic(
-            topic_id, SchemaParser.serialize_patch(operations, many=True)
+            topic_id=topic_id, topic_name=topic_name, operations=SchemaParser.serialize_patch(operations, many=True)
         )